C25Challenge

Documentation for C25Challenge.

C25Challenge.EdgeWeight

C25Challenge.StreetGraph

C25Challenge.Problem

C25Challenge.bestStepLookahead

C25Challenge.calcValue

C25Challenge.compute_distance

C25Challenge.feasible_check

C25Challenge.initStreetGraph

C25Challenge.outInformation

C25Challenge.parallel_solver

C25Challenge.random_parallel_solver

C25Challenge.series_solver

C25Challenge.updateNVisited!

C25Challenge.EdgeWeight — Type

EdgeWeight

Stores the weight of an edge in a StreetGraph

Fields

duration::Integer : The duration to traverse the edge or street

rate::Float64: The rate of taking the edge or street

nvisited::Float64: The number of times the edge has been visited

C25Challenge.StreetGraph — Type

StreetGraph

Master graph representing the street system of a City, wrapper for ValOutDiGraph and stores other properties of the problem.

Fields

graph::SimpleValueGraphs.ValOutDiGraph: Graph Representation of the City

N::Integer: number of cars

start::Integer: The index of the initial junction

totalTime::Integer: time limit of the problem

C25Challenge.Problem — Method

Problem()

Return a StreetGraph

C25Challenge.bestStepLookahead — Function

bestStepLookahead(graph::StreetGraph, currentNode::Integer, depth=5)

A lookahead function that calulates a heuristic and returns the corresponding next best node.

Parameters

graph: The graph format of the City

currentNode: The current junction or node we are looking ahead after

depth: The depth of the lookahead search

C25Challenge.calcValue — Method

calcValue(weight::EdgeWeight)

Calculates a heuristic given an edge.

Parameters

weight: The EdgeWeight that you calculate the heuristic on

C25Challenge.compute_distance — Method

compute_distance(solution::Solution)

Computes the total distance of all itineraries in solution based on the StreetGraph Streets visited several times are only counted once.

C25Challenge.feasible_check — Method

feasible_check(solution, city[; verbose=false])

Check if solution satisfies the constraints of the problme statement The following criteria are considered:

the number of itineraries has to match the number of cars that traverse the city`

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

C25Challenge.initStreetGraph — Method

Used to construct a StreetGraph, returns StreetGraph object

C25Challenge.outInformation — Method

outInformation(g::StreetGraph, node::Integer)

Returns the information about the node. Specifically, the outedgevals, outneighbors, and max rate index outvals, outneighbors, idx_max = outInformation(g::StreetGraph, node::Integer)

Parameters

streetGraph: The graph format of the City

node: The node on the graph format to get information about

C25Challenge.parallel_solver — Function

parallel_solver(streetGraph::StreetGraph, lookahead=5)

A greedy algorithm that solves the problem by sending the cars all at once

Parameters

streetGraph: The graph format of the City

lookahead: The number of steps to lookahead

C25Challenge.random_parallel_solver — Function

random_parallel_solver(streetGraph::StreetGraph, lookahead=5, pct_chance=5)

A greedy algorithm that solves the problem in parallel but with random chance of going again

Parameters

streetGraph: The graph format of the City

lookahead: The number of steps to lookahead

pct_chance: Drives random behavior for a car to go again

C25Challenge.series_solver — Function

series_solver(streetGraph::StreetGraph, lookahead=5)

A greedy algorithm that solves the problem by sending the cars one after another.

Parameters

streetGraph: The graph format of the City

lookahead: The number of steps to lookahead

C25Challenge.updateNVisited! — Method

NEW UTILS !!! NEED TO DOCUMENT THESE